# share/api.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "api.proto",
        "version": "1"
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/share.service.v1.share/DeleteGameArena": {
            "get": {
                "summary": "/share.service.v1.share/DeleteGameArena",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.share.service.v1.DeleteGameArenaResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.share.service.v1.DeleteGameArenaReq"
                        }
                    }
                ],
                "tags": [
                    "share.service.v1.share"
                ]
            }
        },
        "/share.service.v1.share/HListAdd": {
            "get": {
                "summary": "/share.service.v1.share/HListAdd",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.share.service.v1.CommonResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Item",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Key",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ListName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "GameId",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "share.service.v1.share"
                ]
            }
        },
        "/share.service.v1.share/HListAddLast": {
            "get": {
                "summary": "/share.service.v1.share/HListAddLast",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.share.service.v1.CommonResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Item",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Key",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ListName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "GameId",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "share.service.v1.share"
                ]
            }
        },
        "/share.service.v1.share/HListDel": {
            "get": {
                "summary": "/share.service.v1.share/HListDel",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.share.service.v1.CommonResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Item",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Key",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ListName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "GameId",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "share.service.v1.share"
                ]
            }
        },
        "/share.service.v1.share/HListDelAll": {
            "get": {
                "summary": "/share.service.v1.share/HListDelAll",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.share.service.v1.CommonResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Item",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Key",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ListName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "GameId",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "share.service.v1.share"
                ]
            }
        },
        "/share.service.v1.share/HListLoadAll": {
            "get": {
                "summary": "/share.service.v1.share/HListLoadAll",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.share.service.v1.HListLoadResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "Item",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "Key",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "ListName",
                        "in": "query",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "GameId",
                        "in": "query",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "tags": [
                    "share.service.v1.share"
                ]
            }
        },
...
        "/share.service.v1.share/HSetAdd": {
            "get": {
                "summary": "/share.service.v1.share/HSetAdd",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.share.service.v1.CommonResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.share.service.v1.HSetAddReq"
                        }
                    }
                ],
                "tags": [
                    "share.service.v1.share"
                ]
            }
        },
...
    },
    "definitions": {
...
        ".share.service.v1.CommonResp": {
            "type": "object",
            "properties": {},
            "description": "/////////////////////////////////////////"
        },
        ".share.service.v1.DeleteGameArenaReq": {
            "type": "object",
            "properties": {
                "GameArena": {
                    "$ref": "#/definitions/.share.service.v1.GameArena"
                }
            }
        },
        ".share.service.v1.DeleteGameArenaResp": {
            "type": "object",
            "properties": {}
        },
        ".share.service.v1.GameArena": {
            "type": "object",
            "properties": {
                "GameId": {
                    "type": "integer"
                },
                "GroupId": {
                    "type": "integer"
                },
                "ArenaInfo": {
                    "type": "string"
                },
                "ExtraData": {
                    "type": "string"
                },
                "MatchId": {
                    "type": "integer"
                }
            }
        },
...
        ".share.service.v1.HListLoadResp": {
            "type": "object",
            "properties": {
                "Items": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        ".share.service.v1.HSetAddReq": {
            "type": "object",
            "properties": {
                "Item": {
                    "$ref": "#/definitions/.share.service.v1.KeyValue"
                },
                "HashName": {
                    "type": "string"
                },
                "GameId": {
                    "type": "integer"
                }
            }
        },
...
    }
}